refactor: ensureNotNullishをtype/utility.tsに移す
