+++ src/components/Menu/MenuBar/useCommonMenuBarData.ts
@@ -1,7 +1,7 @@
 import { computed, ref, watch } from "vue";
 import { MenuItemData } from "../type";
 import { MaybeComputedMenuBarContent } from "./menuBarData";
-import { ensureNotNullish } from "@/helpers/errorHelper";
+import { ensureNotNullish } from "@/type/utility";
 import { useHotkeyManager, HotkeyAction } from "@/plugins/hotkeyPlugin";
 import { Store } from "@/store";
 import { isProduction } from "@/helpers/platform";
