+++ src/helpers/errorHelper.ts
@@ -1,14 +1,3 @@
-/** 入力がnullかundefinedの場合エラーを投げ、それ以外の場合は入力をそのまま返す */
-export const ensureNotNullish = <T>(
-  value: T | null | undefined,
-  message = "Unexpected nullish value",
-): T => {
-  if (value == null) {
-    throw new Error(message);
-  }
-  return value;
-};
-
 /**
  * ユーザーに表示するメッセージを持つエラー。
  * {@link errorToMessages}でユーザー向けのメッセージとして扱われる。
