+++ src/type/utility.ts
@@ -19,6 +19,17 @@ export function assertNonNullable<T>(
   }
 }
 
+/** 入力がnullかundefinedの場合エラーを投げ、それ以外の場合は入力をそのまま返す */
+export const ensureNotNullish = <T>(
+  value: T | null | undefined,
+  message = "Unexpected nullish value",
+): T => {
+  if (value == null) {
+    throw new Error(message);
+  }
+  return value;
+};
+
 /**
  * never型になるはずの値を使って型システムで非到達をチェックできる便利クラス。
  * valueに指定した値がnever型じゃなかった場合に型システムがエラーを出してくれる。
