+++ tests/e2e/browser/song/ルーラーループ.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from "@playwright/test";
 
 import { gotoHome, navigateToSong } from "../../navigators";
-import { ensureNotNullish } from "@/helpers/errorHelper";
+import { ensureNotNullish } from "@/type/utility";
 
 test.beforeEach(gotoHome);
