+++ tests/e2e/browser/song/歌詞.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect, Page, Locator } from "@playwright/test";
 
 import { gotoHome, navigateToSong } from "../../navigators";
-import { ensureNotNullish } from "@/helpers/errorHelper";
+import { ensureNotNullish } from "@/type/utility";
 
 test.beforeEach(gotoHome);
